refactor traffic logic
